Removal of /rwbyg/ discord server link and legacy pastebin
After careful consideration and polling of the /rwbyg/ members, I've decided to remove the link for the /rwbyg/ discord server (https://discord.gg/CYDQUDZZUJ) from the main page. This is for mainly two reasons:
1. It takes away discussion from the thread. This is the biggest reason driving this change. There is no doubt about it that Discord is a far easier platform to use than 4chan. Additionally, with the mods they have, they are able to remove any bad actors or people they don't like. This instantly makes the Discord a far more attractive place to talk.
2. The residents of the Discord do not represent the current members. There is no crossposting of leaked material or news; there is no thread interaction (except for Kali with his copypasta back in February, but that's the last I've been able to confirm). The two are separate. Platypus even left the general for the Discord. I understand that these Discord members likely come whenever we get on /co/ while RWBY is airing, but our majority of the time is in /trash/, and thus is our main userbase. They have clearly fostered their own community far removed from ours, with their own regulars and multiple topics of discussion beyond RWBY. They may claim they are /rwbyg/, but to be frank, they are merely oldheads that no longer participate where they once came. This is further confirmed by Discord users advertising their Discord in RWBY threads on /a/ or /v/ rather than the general, which I have personally seen.

This may come as a surprise and uncalled for regardless of my reasons, but I assure you, it isn't. /hg/, Halo General, also had this exact same issue. They had a Discord server, but were forced to remove the link because it would drive discussion away from the threads.

I hope any Discord members reading this will not come to the conclusion that I did this because I dislike the server. I do dislike the server, however, my reasoning is based in logic. I wouldn't have made this change if I didn't have reason to. Though, me disliking the server does make this change much easier.

Secondly, the pastebin. This was done without polling the general, but I can't see a reason for opposition other than losing the Discord link (which you can look up and it's the first result in Google). Frankly, it no longer offers anything of real value to today's members. It hasn't been updated since Jan 15, 2021 as of today. It doesn't properly specify what's what, and it doesn't keep up with the general. It holds no value. I thank Samin for his efforts up till this point, and hope I can still rely on him for usage in the MEGA until I can get my torrent box up.