Maintenance update Final update to the site until Viz reuploads RWBY content on their site
- Change V9 animatic link to shitty youtube mirror since RT died
- Removed the archive page off of the main page (it's still on the RT page)
- Removed Beaconship
Beaconship is dead. For good.
http://104.131.69.78/ was the link, but it no longer works.
If anyone has an archive of it, please send it to me or upload it to archive.org.
- Removed screencap archive
Screencap Archive (was a google drive link) is now restricted.
- Updated my RWBY folder
My RWBY folder has received its final update. Just added more images, all unorganized. If you wanted to grab it, now's the time, as I won't be keeping any backups.
- Updated dead links (RIP aniwave)
- Removed unneeded HardcoreMinigolf MEGA since RT Archive exists
- Minor formatting and consistency fixes